// src/app/interfaces/IData.ts
export interface IData {
  economy: {
    priceOfOneKilometer: number;
    maxWeightOfLuggage: number;
    priceOfLuggage: number;
    sale: number | null
  };
  advanced: {
    priceOfOneKilometer: number;
    maxWeightOfLuggage: number;
    priceOfLuggage: number;
    sale: number | null
  };
  luxury: {
    priceOfOneKilometer: number;
    maxWeightOfLuggage: number;
    priceOfLuggage: number;
    sale: number | null
  };
}